+++ lib/webspicy/support/world.rb
@@ -2,15 +2,16 @@ module Webspicy
   module Support
     class World < OpenStruct
 
-      def initialize(folder)
+      def initialize(folder, config = nil)
         super({})
         @folder = folder
+        @config = config
       end
-      attr_reader :folder
+      attr_reader :folder, :config
 
       def method_missing(name, *args, &bl)
         return super if name =~ /=$/
-        file = ['json', 'yml', 'yaml']
+        file = ['json', 'yml', 'yaml', 'rb']
           .map{|e| folder/"#{name}.#{e}" }
           .find{|f| f.file? }
         data = case file && file.ext
@@ -18,6 +19,10 @@ module Webspicy
           JSON.parse(file.read, object_class: OpenStruct)
         when /ya?ml/
           JSON.parse(file.load.to_json, object_class: OpenStruct)
+        when /rb/
+          ::Kernel.eval(file.read).tap{|x|
+            x.config = self.config if x.is_a?(Item)
+          }
         end
         self.send(:"#{name}=", data)
       end
@@ -32,6 +37,10 @@ module Webspicy
           which
         end
       end
+
+      module Item
+        attr_accessor :config
+      end # module Item
 
     end # class World
   end # module Support