+++ spec/unit/support/world/fixtures/queue.rb
@@ -1,0 +1,1 @@
+Webspicy::SpecTests::Queue.new