+++ spec/unit/support/world/test_world.rb
@@ -1,11 +1,24 @@
 require 'spec_helper'
 
 module Webspicy
+  module SpecTests
+    class Queue
+      include Webspicy::Support::World::Item
+
+    end
+  end
+end
+
+module Webspicy
   module Support
     describe World do
 
+      let(:config){
+        Configuration.new(restful_folder)
+      }
+
       subject do
-        World.new Path.dir/"fixtures"
+        World.new Path.dir/"fixtures", config
       end
 
       it 'supports single objects' do
@@ -31,6 +44,12 @@ module Webspicy
       it 'allows setting new attributes on existing objects' do
         subject.single.foo = "bar"
         expect(subject.single.foo).to eql("bar")
+      end
+
+      it 'supports ruby files and evaluates them' do
+        q = subject.queue
+        expect(q).to be_a(Webspicy::SpecTests::Queue)
+        expect(q.config).to be(config)
       end
     end
   end